+++ pcdl/__init__.py
@@ -1,5 +1,4 @@
-from pcdl.pyAnnData import TimeStep, TimeSeries, scaler
-from pcdl.pyMCDS import pyMCDS, graphfile_parser
-from pcdl.pyMCDSts import pyMCDSts, make_gif, make_movie
+from pcdl.timestep import TimeStep, graphfile_parser, scaler
+from pcdl.timeseries import TimeSeries, make_gif, make_movie
 from pcdl.VERSION import __version__
-from pcdl.data_timeseries import install_data, uninstall_data
+from pcdl.output_data import install_data, uninstall_data
