+++ pcdl/timeseries.py
@@ -1,5 +1,5 @@
 #########
-# title: pyMCDSts.py
+# title: timeseries.py
 #
 # language: python3
 # date: 2022-08-22
@@ -7,19 +7,16 @@
 # authors: Patrick Wall, Randy Heiland, Paul Macklin, Elmar Bucher
 #
 # description:
-#     pyMCDSts.py defines an object class, able to load and access
+#     timeseries.py defines an object class, able to load and access
 #     within python a time series of mcds objects loaded from a single
-#     PhysiCell model output directory. pyMCDSts.py was first forked from
+#     PhysiCell model output directory. timeseries.py was first forked from
 #     PhysiCell-Tools python-loader, where it was implemented as
 #     pyMCDS_timeseries.py, then totally rewritten and further developed.
-#
-#     the make_image and make_movie functions are cloned from the PhysiCell
-#     Makefile. note on difference image magick convert and mogrify:
-#     + https://graphicsmagick-tools.narkive.com/9Sowc4HF/gm-tools-mogrify-vs-convert
-#########
+####
 
 
 # load libraries
+import anndata as ad
 import bioio_base
 from bioio.writers import OmeTiffWriter
 import glob
@@ -27,7 +24,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import os
 import pandas as pd
-from pcdl.pyMCDS import pyMCDS, es_coor_cell, es_coor_conc
+from pcdl.timestep import TimeStep, es_coor_cell, es_coor_conc, _anndextract
 import platform
 import sys
 import xml.etree.ElementTree as etree
@@ -161,7 +158,7 @@ def make_movie(path, interface='jpeg', framerate=12):
 # classes #
 ###########
 
-class pyMCDSts:
+class TimeSeries:
     def __init__(self, output_path='.', custom_data_type={}, load=True, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True):
         """
         input:
@@ -185,7 +182,7 @@ class pyMCDSts:
                 processing, similar to the original pyMCDS_cells.py script.
 
             graph: boole; default True
-                should the graphs be loaded?
+                should the graphs, like cell_neighbor_graph.txt, be loaded?
                 setting graph to False will use less memory and speed up processing.
 
             physiboss: boole; default True
@@ -202,12 +199,12 @@ class pyMCDSts:
                 setting verbose to False for less text output, while processing.
 
         output:
-            mcdsts: pyMCDSts class instance
+            mcdsts: TimeSeries class instance
                 this instance offers functions to process all stored time steps
                 from a simulation.
 
         description:
-            pyMCDSts.__init__ generates a class instance the instance offers
+            TimeSeries.__init__ generates a class instance the instance offers
             functions to process all time steps in the output_path directory.
         """
         output_path = output_path.replace('\\','/')
@@ -216,7 +213,7 @@ class pyMCDSts:
         if (output_path.endswith('/')) and (len(output_path) > 1):
             output_path = output_path[:-1]
         if not os.path.isdir(output_path):
-            print(f'Error @ pyMCDSts.__init__ : this is not a path! could not load {output_path}.')
+            print(f'Error @ TimeSeries.__init__ : this is not a path! could not load {output_path}.')
         self.path = output_path
         self.ls_xmlfile = [s_pathfile.replace('\\','/').split('/')[-1] for s_pathfile in sorted(glob.glob(self.path + f'/output*.xml'))]  # bue 2022-10-22: is output*.xml always the correct pattern?
         self.custom_data_type = custom_data_type
@@ -229,6 +226,7 @@ class pyMCDSts:
             self.read_mcds()
         else:
             self.l_mcds = None
+        self.l_annmcds = None
 
 
     def set_verbose_false(self):
@@ -278,7 +276,7 @@ class pyMCDSts:
     def get_xmlfile_list(self):
         """
         input:
-            self: pyMCDSts class instance.
+            self: TimeSeries class instance.
 
         output:
             xmlfile_list: list of strings
@@ -296,7 +294,7 @@ class pyMCDSts:
     def read_mcds(self, xmlfile_list=None):
         """
         input:
-            self: pyMCDSts class instance.
+            self: TimeSeries class instance.
 
             xmlfile_list: list of strings; default None
                 list of physicell output*.xml strings.
@@ -306,7 +304,7 @@ class pyMCDSts:
 
         description:
             the function returns a list of mcds objects loaded by
-            pyMCDS calls.
+            TimeStep calls.
         """
         # handle input
         if (xmlfile_list is None):
@@ -317,7 +315,7 @@ class pyMCDSts:
         # load mcds objects into list
         l_mcds = []
         for s_xmlpathfile in ls_xmlpathfile:
-            mcds = pyMCDS(
+            mcds = TimeStep(
                 xmlfile = s_xmlpathfile,
                 custom_data_type = self.custom_data_type,
                 microenv = self.microenv,
@@ -339,7 +337,7 @@ class pyMCDSts:
     def get_mcds_list(self):
         """
         input:
-            self: pyMCDSts class instance.
+            self: TimeSeries class instance.
 
         output:
             self.l_mcds: list of chronologically ordered mcds objects.
@@ -357,7 +355,7 @@ class pyMCDSts:
     def get_conc_df(self, values=1, drop=set(), keep=set(), collapse=True):
         """
         input:
-            self: pyMCDSts class instance.
+            self: TimeSeries class instance.
 
             values: integer; default is 1
                 minimal number of values a variable has to have to be outputted.
@@ -442,7 +440,7 @@ class pyMCDSts:
     def get_conc_attribute(self, values=1, drop=set(), keep=set(), allvalues=False):
         """
         input:
-            self: pyMCDSts class instance.
+            self: TimeSeries class instance.
 
             values: integer; default is 1
                 minimal number of values a variable has to have
@@ -508,7 +506,7 @@ class pyMCDSts:
     def plot_contour(self, focus, z_slice=0.0, extrema=None, alpha=1, fill=True, cmap='viridis', title='', grid=True, xlim=None, ylim=None, xyequal=True, figsizepx=None, ext='jpeg', figbgcolor=None):
         """
         input:
-            self: pyMCDSts class instance
+            self: TimeSeries class instance
 
             focus: string
                 column name within conc dataframe, for example.
@@ -678,7 +676,7 @@ class pyMCDSts:
     def get_cell_df(self, values=1, drop=set(), keep=set(), collapse=True):
         """
         input:
-            self: pyMCDSts class instance.
+            self: TimeSeries class instance.
 
             values: integer; default is 1
                 minimal number of values a variable has to have to be outputted.
@@ -766,7 +764,7 @@ class pyMCDSts:
     def get_cell_attribute(self, values=1, drop=set(), keep=set(), allvalues=False):
         """
         input:
-            self: pyMCDSts class instance.
+            self: TimeSeries class instance.
 
             values: integer; default is 1
                 minimal number of values a variable has to have
@@ -834,7 +832,7 @@ class pyMCDSts:
     def plot_scatter(self, focus='cell_type', z_slice=0.0, z_axis=None, alpha=1, cmap='viridis', title='', grid=True, legend_loc='lower left', xlim=None, ylim=None, xyequal=True, s=None, figsizepx=None, ext='jpeg', figbgcolor=None):
         """
         input:
-            self: pyMCDSts class instance
+            self: TimeSeries class instance
 
             focus: string; default is 'cell_type'
                 column name within cell dataframe.
@@ -1126,7 +1124,7 @@ class pyMCDSts:
     def plot_timeseries(self, focus_cat=None, focus_num=None, aggregate_num=np.nanmean, frame='cell', z_slice=None, logy=False, ylim=None, secondary_y=None, subplots=False, sharex=False, sharey=False, linestyle='-', linewidth=None, cmap=None, color=None, grid=True, legend=True, yunit=None, title=None, ax=None, figsizepx=[640, 480], ext=None, figbgcolor=None):
         """
         input:
-            self: pyMCDSts class instance
+            self: TimeSeries class instance
 
             focus_cat: string; default is None
                 categorical or boolean data column within dataframe specified under frame.
@@ -1291,7 +1289,7 @@ class pyMCDSts:
                 mcds.set_verbose_true()
             # error
             else:
-                sys.exit(f"Error @ pyMCDSts.plot_timeseries : unknown frame {frame}. known are cell_df and conc_df.")
+                sys.exit(f"Error @ TimeSeries.plot_timeseries : unknown frame {frame}. known are cell_df and conc_df.")
             # handle z_slize
             if not (z_slice is None):
                 df_frame = df_frame.loc[(df_frame.mesh_center_p == z_slice),:]
@@ -1306,7 +1304,7 @@ class pyMCDSts:
                 df_aggregate = o_aggregate.loc[:,[focus_num]]
                 df_aggregate.columns = [r_time]
             else:
-                sys.exit(f'Error @ pyMCDSts.plot_timeseries : {aggregate_num} calculation returns unexpected variable type {type(o_aggregate)}.\nthe expected type is a pandas Series or DataFrame.')
+                sys.exit(f'Error @ TimeSeries.plot_timeseries : {aggregate_num} calculation returns unexpected variable type {type(o_aggregate)}.\nthe expected type is a pandas Series or DataFrame.')
             # store result
             if (df_series is None):
                 df_series = df_aggregate
@@ -1407,7 +1405,7 @@ class pyMCDSts:
     def make_graph_gml(self, graph_type, edge_attribute=True, node_attribute=[]):
         """
         input:
-            self: pyMCDS class instance.
+            self: TimeSeries class instance.
 
             graph_type: string
                 to specify which physicell output data should be processed.
@@ -1454,3 +1452,170 @@ class pyMCDSts:
 
         # outout
         return ls_pathfile
+
+
+    ## ANNDATA RELATED FUNCTIONS ##
+
+    def get_anndata(self, values=1, drop=set(), keep=set(), scale='maxabs', collapse=True, keep_mcds=True):
+        """
+        input:
+            values: integer; default is 1
+                minimal number of values a variable has to have to be outputted.
+                variables that have only 1 state carry no information.
+                None is a state too.
+
+            drop: set of strings; default is an empty set
+                set of column labels to be dropped for the dataframe.
+                don't worry: essential columns like ID, coordinates
+                and time will never be dropped.
+                Attention: when the keep parameter is given, then
+                the drop parameter has to be an empty set!
+
+            keep: set of strings; default is an empty set
+                set of column labels to be kept in the dataframe.
+                don't worry: essential columns like ID, coordinates
+                and time will always be kept.
+
+            scale: string; default 'maxabs'
+                specify how the data should be scaled.
+                possible values are None, maxabs, minmax, std.
+                for more input, check out: help(pcdl.scaler)
+
+            collapse: boole; default True
+                should all mcds time steps from the time series be collapsed
+                into one single anndata object, or a list of anndata objects
+                for each time step?
+
+            keep_mcds: boole; default True
+                should the loaded original mcds be kept in memory
+                after transformation?
+
+        output:
+            annmcds or self.l_annmcds: anndata object or list of anndata objects.
+                what is returned depends on the collapse setting.
+
+        description:
+            function to transform mcds time steps into one or many
+            anndata objects for downstream analysis.
+        """
+        # initialize vaiable
+        l_annmcds = []
+        df_anncount = None
+        df_annobs = None
+        ar_annobsm = None
+
+        # variable triage
+        if (values < 2):
+            ls_column = list(self.l_mcds[0].get_cell_df(drop=drop, keep=keep).columns)
+        else:
+            ls_column = sorted(es_coor_cell.difference({'ID'}))
+            ls_column.extend(sorted(self.get_cell_attribute(values=values, drop=drop, keep=keep, allvalues=False).keys()))
+
+        # collapse warning
+        if collapse and self.verbose:
+            print('Warning @ mcdsts.get_anndata : only df_cell data, but not graph data, can be collapsed.')
+
+        # processing
+        lann_mcds = []
+        i_mcds = len(self.l_mcds)
+        for i in range(i_mcds):
+            # fetch mcds
+            if keep_mcds:
+                mcds = self.l_mcds[i]
+            else:
+                mcds = self.l_mcds.pop(0)
+            # extract physicell version
+            s_physicellv = mcds.get_physicell_version(),
+            # extract time and dataframes
+            r_time = round(mcds.get_time(),9)
+            if self.verbose:
+                print(f'processing: {i+1}/{i_mcds} {r_time}[min] mcds into anndata obj.')
+            df_cell = mcds.get_cell_df()
+            df_cell = df_cell.loc[:,ls_column]
+
+            # pack collapsed
+            if collapse:
+                # extract
+                df_count, df_obs, d_obsm, d_obsp, d_uns = _anndextract(
+                    df_cell=df_cell,
+                    scale = scale,
+                    #graph_attached = {},
+                    #graph_neighbor = {},
+                    #graph_spring = {},
+                    #graph_method = s_physicellv,
+                )
+                # count
+                df_count.reset_index(inplace=True)
+                df_count.index = df_count.ID + f'id_{r_time}min'
+                df_count.index.name = 'id_time'
+                df_count.drop('ID', axis=1, inplace=True)
+                if df_anncount is None:
+                    df_anncount = df_count
+                else:
+                    df_anncount = pd.concat([df_anncount, df_count], axis=0)
+                # obs
+                df_obs.reset_index(inplace=True)
+                df_obs.index = df_obs.ID + f'id_{r_time}min'
+                df_obs.index.name = 'id_time'
+                if df_annobs is None:
+                    df_annobs = df_obs
+                else:
+                    df_annobs = pd.concat([df_annobs, df_obs], axis=0)
+                # obsm (spatial)
+                if ar_annobsm is None:
+                    ar_annobsm = d_obsm['spatial']
+                else:
+                    ar_annobsm = np.vstack([ar_annobsm, d_obsm['spatial']])
+                # obsp: nop (graph)
+                # uns: nop (graph)
+
+            # pack not collapsed
+            else:
+                # extract
+                df_count, df_obs, d_obsm, d_obsp, d_uns = _anndextract(
+                    df_cell=df_cell,
+                    scale = scale,
+                    graph_attached = mcds.get_attached_graph_dict(),
+                    graph_neighbor = mcds.get_neighbor_graph_dict(),
+                    graph_spring = mcds.get_spring_graph_dict(),
+                    graph_method = s_physicellv,
+                )
+                # annmcds
+                ann_mcds = ad.AnnData(
+                    X = df_count,
+                    obs = df_obs,
+                    obsm = d_obsm,
+                    obsp = d_obsp,
+                    uns = d_uns,
+                )
+                lann_mcds.append(ann_mcds)
+
+        # output
+        if collapse:
+            ann_mcdsts = ad.AnnData(
+                X = df_anncount,
+                obs = df_annobs,
+                obsm = {'spatial': ar_annobsm},
+                #obsp = d_obsp,
+                #uns = d_uns
+            )
+            return ann_mcdsts
+        else:
+            self.l_annmcds = lann_mcds
+            return self.l_annmcds
+
+
+    def get_annmcds_list(self):
+        """
+        input:
+            self: TimeSeries class instance.
+
+        output:
+            self.l_annmcds: list of chronologically ordered anndata mcds objects.
+                watch out, this is a pointer to the
+                self.l_annmcds list of anndata mcds objects, not a copy of self.l_annmcds!
+
+        description:
+            function returns a binding to the self.l_annmcds list of anndata mcds objects.
+        """
+        return self.l_annmcds
