+++ pcdl/timestep.py
@@ -1,5 +1,5 @@
 #########
-# title: pyMCDS.py
+# title: timestep.py
 #
 # language: python3
 # date: 2022-08-22
@@ -7,14 +7,15 @@
 # authors: Patrick Wall, Randy Heiland, Furkan Kurtoglu, Paul Macklin, Elmar Bucher
 #
 # description:
-#     pyMCDS.py definds an object class, able to load and access
+#     timestep.py definds an object class, able to load and access
 #     within python a single time step from the PhysiCell model output folder.
-#     pyMCDS.py was forked from the original PhysiCell-Tools python-loader
+#     timestep.py was forked from the original PhysiCell-Tools python-loader
 #     implementation and further developed.
 #########
 
 
 # load library
+import anndata as ad
 import bioio_base
 from bioio.writers import OmeTiffWriter
 import matplotlib.pyplot as plt
@@ -26,9 +27,11 @@ import pandas as pd
 from pcdl import imagine
 from pcdl import pdplt
 from scipy import io
+from scipy import sparse
 import sys
 import vtk
 from vtkmodules.vtkCommonCore import vtkPoints
+import warnings
 import xml.etree.ElementTree as etree
 from pcdl.VERSION import __version__
 
@@ -201,8 +204,239 @@ def graphfile_parser(s_pathfile):
     return dei_graph
 
 
+def scaler(df_x, scale='maxabs'):
+    """
+    input:
+        df_x: pandas dataframe
+              one attribute per column, one sample per row.
+
+        scale: string; default 'maxabs'
+            None: no scaling. set scale to None if you would like to have
+                raw data or scale, transform, and normalize the data later.
+
+            maxabs: maximum absolute value distance scaler will linearly map
+                all values into a [-1, 1] interval. if the original data
+                has no negative values, the result will be the same as with
+                the minmax scaler (except with attributes with only one value).
+                if the attribute has only zeros, the value will be set to 0.
+
+            minmax: minimum maximum distance scaler will map all values
+                linearly into a [0, 1] interval.
+                if the attribute has only one value, the value will be set to 0.
+
+            std: standard deviation scaler will result in sigmas.
+                each attribute will be mean centered around 0.
+                ddof delta degree of freedom is set to 1 because it is assumed
+                that the values are samples out of the population
+                and not the entire population. it is incomprehensible to me
+                that the equivalent sklearn method has ddof set to 0.
+                if the attribute has only one value, the value will be set to 0.
+
+    output:
+        df_x: pandas dataframe
+            scaled df_x dataframe.
+
+    description:
+        inspired by scikit-learn's preprocessing scaling method, this function
+        offers a re-implementation of the linear re-scaling methods maxabs,
+        minmax, and scale.
+
+        the robust scaler methods (quantile based) found in scikit-learn are
+        missing. since we deal with simulated data, we don't expect heavy
+        outliers, and if they exist, then they are of interest.
+        the power and quantile based transformation methods and unit circle
+        based normalizer methods found there are missing too.
+        if you need to apply any such methods, you can do so to an anndata object
+        like this:
+
+        from sklearn import preprocessing
+        adata.obsm["X_scaled"] = preprocessing.scale(adata.X)
+
+    + https://scikit-learn.org/stable/auto_examples/preprocessing/plot_all_scaling.html
+    + https://scikit-learn.org/stable/modules/classes.html#module-sklearn.preprocessing
+    + https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.maxabs_scale.html
+    + https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.minmax_scale.html
+    + https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.scale.html
+    """
+    if scale is None:
+        pass
+    # -1,1
+    elif scale == 'maxabs':
+        a_x = df_x.values
+        warnings.filterwarnings('ignore', category=RuntimeWarning)
+        a_maxabs = a_x / abs(a_x).max(axis=0)
+        warnings.simplefilter('default')
+        a_maxabs[np.isnan(a_maxabs)] = 0  # fix if entier column is 0
+        df_x = pd.DataFrame(a_maxabs, columns=df_x.columns, index=df_x.index)
+    # 0,1
+    elif scale == 'minmax':
+        a_x = df_x.values
+        warnings.simplefilter("ignore")
+        warnings.filterwarnings('ignore', category=RuntimeWarning)
+        a_minmax = (a_x - a_x.min(axis=0)) / (a_x.max(axis=0) - a_x.min(axis=0))
+        warnings.simplefilter('default')
+        a_minmax[np.isnan(a_minmax)] = 0  # fix if entier column has same value
+        df_x = pd.DataFrame(a_minmax, columns=df_x.columns, index=df_x.index)
+    # sigma
+    elif scale == 'std':
+        a_x = df_x.values
+        warnings.filterwarnings('ignore', category=RuntimeWarning)
+        a_std = (a_x - a_x.mean(axis=0)) / a_x.std(axis=0, ddof=1)
+        warnings.simplefilter('default')
+        a_std[np.isnan(a_std)] = 0  # fix if entier column has same value
+        df_x = pd.DataFrame(a_std, columns=df_x.columns, index=df_x.index)
+    else:
+        raise ValueError(f"Error @ scaler : unknown scale algorithm {scale} detected. known are [None, 'maxabs', 'minmax', 'std'].")
+
+    return df_x
+
+
+def _anndextract(df_cell, scale='maxabs', graph_attached={}, graph_neighbor={}, graph_spring={}, graph_method='PhysiCell'):
+    """
+    input:
+        df_cell:  pandas dataframe
+            data frame retrieved with the mcds.get_cell_df function.
+
+        scale: string; default maxabs
+            specify how the data should be scaled.
+            possible values are None, maxabs, minmax, std.
+            for more input, check out: help(pcdl.scaler).
+
+        graph_attached: dict; default {}
+            attached graph dictionary, retrieved with
+            with the mcds.get_attched_graph() function.
+
+        graph_neighbor: dict; default {}
+            neighbor graph dictionary, retrieved
+            with the mcds.get_neighbor_graph() function.
+
+        graph_spring: dict; default {}
+            spring_attached graph dictionary, retrieved
+            with the mcds.get_spring_graph_dict() function.
+
+        graph_method: string; default PhysiCell
+            method how the graphs were generated.
+
+    output:
+        df_count, df_obs, d_obsm, d_obsp, d_uns dataframes and dictionaries,
+            ready to be backed into an anndata object.
+
+    description:
+        this function takes a pcdl df_cell pandas dataframe and re-formats
+        it into a set of two dataframes (df_count, df_obs),
+        two dictionary of numpy array (d_obsm, d_obsp),
+        and one dictionary of string (d_uns),
+        which downstream might be transformed into an anndata object.
+    """
+    # transform index to string
+    df_coor = df_cell.loc[:,['position_x','position_y','position_z']].copy()
+    df_cell.index = df_cell.index.astype(str)
+
+    # build obs anndata object (annotation of observations)
+    df_obs = df_cell.loc[:,['mesh_center_p','time']].copy()
+    df_obs.columns = ['z_layer', 'time']
+
+    # buil obsm anndata object spatial (multi-dimensional annotation of observations)
+    if (len(set(df_cell.position_z)) == 1):
+        df_obsm = df_cell.loc[:,['position_x','position_y']].copy()
+    else:
+        df_obsm = df_cell.loc[:,['position_x','position_y','position_z']].copy()
+    d_obsm = {"spatial": df_obsm.values}
+
+    # build obsp and uns anndata object graph (pairwise annotation of obeservation) and (unstructured data)
+    ####
+    # acknowledgement:
+    #   this code is inspired from the tysserand add_to_AnnData impelmentation
+    #   from Alexis Coullomb form the Pancaldi Lab.
+    #   https://github.com/VeraPancaldiLab/tysserand/blob/main/tysserand/tysserand.py#L1546
+    ####
+    # extract cell_id to index mapping (i always loved perl)
+    di_ididx = df_cell.reset_index().loc[:,'ID'].reset_index().astype(int).set_index('ID').squeeze().to_dict()
+    # transform cell id graph dict to index matrix and pack for anndata
+    d_obsp = {}  # pairwise annotation of obeservation
+    d_uns = {}  # unstructured data
+    for s_graph, dei_graph in [('neighbor', graph_neighbor), ('attached', graph_attached), ('spring', graph_spring)]:
+        lli_edge = []
+        lr_distance = []
+        for i_src, ei_dst in dei_graph.items():
+            for i_dst in ei_dst:
+                # extract edge
+                lli_edge.append([di_ididx[i_src], di_ididx[i_dst]])
+                r_distance = ((df_coor.loc[i_src,:].values -  df_coor.loc[i_dst,:].values)**2).sum()**(1/2)
+                lr_distance.append(r_distance)
+        # if there is a graph
+        if (len(lli_edge) > 0):
+            # handle edge data
+            ai_edge = np.array(lli_edge, dtype=np.uint)
+            # handle connection data
+            ai_conectivity = np.ones(ai_edge.shape[0], dtype=np.uint16)
+            ai_conectivity_sparse = sparse.csr_matrix(
+                (ai_conectivity, (ai_edge[:,0], ai_edge[:,1])),
+                shape = (df_cell.shape[0], df_cell.shape[0]),
+                dtype = np.uint
+            )
+            # handle distance data
+            ar_distance  = np.array(lr_distance, dtype=np.float64)
+            ar_distance_sparse = sparse.csr_matrix(
+                (ar_distance, (ai_edge[:,0], ai_edge[:,1])),
+                shape = (df_cell.shape[0], df_cell.shape[0]),
+                dtype = np.float64
+            )
+            # pack obsp
+            d_obsp.update({
+                f'physicell_{s_graph}_conectivities': ai_conectivity_sparse,
+                f'physicell_{s_graph}_distances': ar_distance_sparse,
+            })
+            # pack uns
+            d_uns.update({
+                s_graph : {
+                    'connectivities_key': f'physicell_{s_graph}_conectivities',
+                    'distances_key': f'physicell_{s_graph}_distances',
+                    'params': {
+                        'metric': 'euclidean',
+                        'method': graph_method,
+                    }
+                }
+            })
+
+    # extract discrete cell data
+    es_drop = set(df_cell.columns).intersection({
+        'voxel_i', 'voxel_j', 'voxel_k',
+        'mesh_center_m', 'mesh_center_n', 'mesh_center_p',
+        'position_x', 'position_y','position_z',
+        'time', 'runtime', 'xmlfile',
+    })
+    df_cell.drop(es_drop, axis=1, inplace=True)  # maybe obs?
+
+    # dectect variable types
+    des_type = {'float': set(), 'int': set(), 'bool': set(), 'str': set()}
+    for _, se_cell in df_cell.items():
+        if str(se_cell.dtype).startswith('float'):
+            des_type['float'].add(se_cell.name)
+        elif str(se_cell.dtype).startswith('int'):
+            des_type['int'].add(se_cell.name)
+        elif str(se_cell.dtype).startswith('bool'):
+            des_type['bool'].add(se_cell.name)
+        elif str(se_cell.dtype).startswith('object'):
+            des_type['str'].add(se_cell.name)
+        else:
+            print(f'Error @ TimeSeries.get_anndata : column {se_cell.name} detected with unknown dtype {str(se_cell.dtype)}.')
+
+    # build on obs and X anndata object
+    df_cat = df_cell.loc[:,sorted(des_type['str'])].copy()
+    df_obs = pd.merge(df_obs, df_cat, left_index=True, right_index=True)
+    es_num = des_type['float'].union(des_type['int'].union(des_type['bool']))
+    df_count = df_cell.loc[:,sorted(es_num)].copy()
+    for s_col in des_type['bool']:
+        df_count[s_col] = df_count[s_col].astype(int)
+    df_count = scaler(df_count, scale=scale)
+
+    # return
+    return(df_count, df_obs, d_obsm, d_obsp, d_uns)
+
+
 # object classes
-class pyMCDS:
+class TimeStep:
     def __init__(self, xmlfile, output_path='.', custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True):
         """
         input:
@@ -226,7 +460,7 @@ class pyMCDS:
                 processing, similar to the original pyMCDS_cells.py script.
 
             graph: boole; default True
-                should the graphs be loaded?
+                should the graphs, like cell_neighbor_graph.txt, be loaded?
                 setting graph to False will use less memory and speed up processing.
 
             physiboss: boole; default True
@@ -243,11 +477,11 @@ class pyMCDS:
                 setting verbose to False for less text output, while processing.
 
         output:
-            mcds: pyMCDS class instance
+            mcds: TimeStep class instance
                 all fetched content is stored at mcds.data.
 
         description:
-            pyMCDS.__init__ will generate a class instance with a
+            TimeStep.__init__ will generate a class instance with a
             dictionary of dictionaries data structure that contains all
             output from a single PhysiCell model time step. furthermore,
             this class, and as such it's instances, offers functions
@@ -268,6 +502,7 @@ class pyMCDS:
         self.verbose = verbose
         self.data = self._read_xml(xmlfile, output_path)
 
+
     def set_verbose_false(self):
         """
         input:
@@ -280,6 +515,7 @@ class pyMCDS:
         """
         self.verbose = False
         #print(f'pcdl: set mcds.verbose = False.')
+
 
     def set_verbose_true(self):
         """
@@ -621,15 +857,15 @@ class pyMCDS:
 
         if (x < tr_x[0]) or (x > tr_x[1]):
             if self.verbose:
-                print(f'Warning @ pyMCDS.is_in_mesh : x = {x} out of bounds: x-range is {tr_x}.')
+                print(f'Warning @ TimeStep.is_in_mesh : x = {x} out of bounds: x-range is {tr_x}.')
             b_isinmesh = False
         elif (y < tr_y[0]) or (y > tr_y[1]):
             if self.verbose:
-                print(f'Warning @ pyMCDS.is_in_mesh : y = {y} out of bounds: y-range is {tr_y}.')
+                print(f'Warning @ TimeStep.is_in_mesh : y = {y} out of bounds: y-range is {tr_y}.')
             b_isinmesh = False
         elif (z < tr_z[0]) or (z > tr_z[1]):
             if self.verbose:
-                print(f'Warning @ pyMCDS.is_in_mesh : z = {z} out of bounds: z-range is {tr_z}.')
+                print(f'Warning @ TimeStep.is_in_mesh : z = {z} out of bounds: z-range is {tr_z}.')
             b_isinmesh = False
 
         # output
@@ -712,7 +948,7 @@ class pyMCDS:
         """
         ar_volume = np.unique(self.data['mesh']['volumes'])
         if ar_volume.shape != (1,):
-            sys.exit(f'Error @ pyMCDS.get_voxel_volume : mesh is not built out of a unique voxel volume {ar_volume}.')
+            sys.exit(f'Error @ TimeStep.get_voxel_volume : mesh is not built out of a unique voxel volume {ar_volume}.')
         r_volume = ar_volume[0]
         return r_volume
 
@@ -861,7 +1097,7 @@ class pyMCDS:
             _, _, ar_p_axis = self.get_mesh_mnp_axis()
             if not (z_slice in ar_p_axis):
                 if self.verbose:
-                    print(f'Warning @ pyMCDS.get_concentration : specified z_slice {z_slice} is not an element of the z-axis mesh centers set {ar_p_axis}.')
+                    print(f'Warning @ TimeStep.get_concentration : specified z_slice {z_slice} is not an element of the z-axis mesh centers set {ar_p_axis}.')
                 if halt:
                     sys.exit('Processing stopped!')
                 else:
@@ -964,14 +1200,14 @@ class pyMCDS:
         """
         # check keep and drop
         if (len(keep) > 0) and (len(drop) > 0):
-            sys.exit(f"Error @ pyMCDS.get_conc_df : when keep is given {keep}, then drop has to be an empty set {drop}!")
+            sys.exit(f"Error @ TimeStep.get_conc_df : when keep is given {keep}, then drop has to be an empty set {drop}!")
 
         # check if z_slice is a mesh center or None
         if not (z_slice is None):
             _, _, ar_p_axis = self.get_mesh_mnp_axis()
             if not (z_slice in ar_p_axis):
                 if self.verbose:
-                    print(f'Warning @ pyMCDS.get_conc_df : specified z_slice {z_slice} is not an element of the z-axis mesh centers set {ar_p_axis}.')
+                    print(f'Warning @ TimeStep.get_conc_df : specified z_slice {z_slice} is not an element of the z-axis mesh centers set {ar_p_axis}.')
                 if halt:
                     sys.exit('Processing stopped!')
                 else:
@@ -1130,7 +1366,7 @@ class pyMCDS:
                 figsizepx = [i_width, i_height]
             except FileNotFoundError:
                 if self.verbose:
-                    print(f'Warning @ pyMCDS.plot_contour : could not load {s_pathfile} to auto detect figsizepx. take default.')
+                    print(f'Warning @ TimeStep.plot_contour : could not load {s_pathfile} to auto detect figsizepx. take default.')
                 figsizepx = [640, 480]
 
         # handle figure size
@@ -1503,7 +1739,7 @@ class pyMCDS:
         """
         # check keep and drop
         if (len(keep) > 0) and (len(drop) > 0):
-            sys.exit(f"Error @ pyMCDS.get_cell_df : when keep is given {keep}, then drop has to be an empty set {drop}!")
+            sys.exit(f"Error @ TimeStep.get_cell_df : when keep is given {keep}, then drop has to be an empty set {drop}!")
 
         # get cell position and more
         df_cell = pd.DataFrame(self.data['discrete_cells']['data'])
@@ -1823,7 +2059,7 @@ class pyMCDS:
                         s = int(round((r_radius)**2))
                     else:
                         if self.verbose:
-                            print(f'Warning @ pyMCDSts.plot_scatter : these agents are not circles.')
+                            print(f'Warning @ TimeStepts.plot_scatter : these agents are not circles.')
                         s = plt.rcParams['lines.markersize']**2
                     if self.verbose:
                         print(f's set to {s}.')
@@ -1833,7 +2069,7 @@ class pyMCDS:
                     figsizepx = [i_width, i_height]
             except FileNotFoundError:
                 if self.verbose:
-                    print(f'Warning @ pyMCDSts.plot_scatter : could not load {s_pathfile}.')
+                    print(f'Warning @ TimeStepts.plot_scatter : could not load {s_pathfile}.')
                 if s is None:
                     s = plt.rcParams['lines.markersize']**2
                     if self.verbose:
@@ -2059,7 +2295,7 @@ class pyMCDS:
             elif (df_cell.loc[:, s_attribute].dtype == int) or (df_cell.loc[:, s_attribute].dtype == float):  # in {int, np.int_, np.int8, np.int16, np.int32, np.int64, float, np.float16, np.float32, np.float64, np.float128}):
                 voa_data = vtk.vtkFloatArray()
             else:
-                sys.exit(f'Error @ pyMCDS.make_cell_vtk : {s_attribute} {df_cell.loc[:, s_attribute].dtype} unknown df_cell column data type.')
+                sys.exit(f'Error @ TimeStep.make_cell_vtk : {s_attribute} {df_cell.loc[:, s_attribute].dtype} unknown df_cell column data type.')
 
             voa_data.SetName(s_attribute)
             for i in df_cell.index:
@@ -2241,9 +2477,9 @@ class pyMCDS:
         # filter and shift
         df_cell = df_cell.loc[:, ls_coor_xyz + ['cell_type', cell_attribute]]
         if (cell_attribute == 'cell_type'):
-            sys.exit(f'Error @ pyMCDS.make_ome_tiff : cell_attribute cannot be cell_type.')
+            sys.exit(f'Error @ TimeStep.make_ome_tiff : cell_attribute cannot be cell_type.')
         elif (df_cell.loc[:, cell_attribute].dtype == str) or (df_cell.loc[:, cell_attribute].dtype == np.object_):  # in {str, np.str_, np.object_}):
-            sys.exit(f'Error @ pyMCDS.make_ome_tiff : {cell_attribute} {df_cell.loc[:, cell_attribute].dtype} cell_attribute cannot be string or object. cell_attribute has to be boolean, integer, or float.')
+            sys.exit(f'Error @ TimeStep.make_ome_tiff : {cell_attribute} {df_cell.loc[:, cell_attribute].dtype} cell_attribute cannot be string or object. cell_attribute has to be boolean, integer, or float.')
         elif (df_cell.loc[:, cell_attribute].dtype == bool): # in {bool, np.bool_, np.bool}):
             df_cell = df_cell.astype({cell_attribute: int})
         df_cell.loc[:, 'position_x'] = (df_cell.loc[:, 'position_x'] - self.get_xyz_range()[0][0]).round()
@@ -2256,7 +2492,7 @@ class pyMCDS:
         # check for duplicates: two cell at exactelly the same xyz position.
         #if self.verbose and df_cell.loc[:,['voxel_x', 'voxel_y', 'voxel_z']].duplicated().any():
         #    df_duplicate = df_cell.loc[(df_cell.loc[:, ['voxel_x', 'voxel_y', 'voxel_z']].duplicated()), :]
-        #    sys.exit(f"Error @ pyMCDS.make_ome_tiff : {df_duplicate} cells at exactely the same xyz voxel position detected. cannot pivot!")
+        #    sys.exit(f"Error @ TimeStep.make_ome_tiff : {df_duplicate} cells at exactely the same xyz voxel position detected. cannot pivot!")
 
         # pivot cell_type
         df_cell = df_cell.pivot_table(index=ls_coor, columns='cell_type', values=cell_attribute, aggfunc='sum').reset_index()  # fill_value is na
@@ -2275,7 +2511,7 @@ class pyMCDS:
             elif s_channel in set(ls_celltype):
                 df_channel = df_cell.loc[:, ls_coor + [s_channel]]
             else:
-                sys.exit(f'Error @ pyMCDS.make_ome_tiff : {s_channel} unknown channel detected. not in substrate and cell type list {ls_substrate} {ls_celltype}!')
+                sys.exit(f'Error @ TimeStep.make_ome_tiff : {s_channel} unknown channel detected. not in substrate and cell type list {ls_substrate} {ls_celltype}!')
 
             # each z axis
             la_zyx_img = []
@@ -2475,7 +2711,7 @@ class pyMCDS:
                 elif (o_attribute.dtype == float):  #in {float, np.float16, np.float32, np.float64, np.float128}):
                     f.write(f'    {s_attribute} {o_attribute}\n')
                 else:
-                    sys.exit(f'Error @ pyMCDS.make_graph_gml : attribute {o_attribute}; type {o_attribute.dtype}; type seems not to be bool, int, float, or string.')
+                    sys.exit(f'Error @ TimeStep.make_graph_gml : attribute {o_attribute}; type {o_attribute.dtype}; type seems not to be bool, int, float, or string.')
             f.write(f'  ]\n')
             # edge
             for i_dst in ei_dst:
@@ -2500,12 +2736,71 @@ class pyMCDS:
         return s_gmlpathfile
 
 
+    ## ANNDATA RELATED FUNCTIONS ##
+
+    def get_anndata(self, values=1, drop=set(), keep=set(), scale='maxabs'):
+        """
+        input:
+            values: integer; default is 1
+                minimal number of values a variable has to have to be outputted.
+                variables that have only 1 state carry no information.
+                None is a state too.
+
+            drop: set of strings; default is an empty set
+                set of column labels to be dropped for the dataframe.
+                don't worry: essential columns like ID, coordinates
+                and time will never be dropped.
+                Attention: when the keep parameter is given, then
+                the drop parameter has to be an empty set!
+
+            keep: set of strings; default is an empty set
+                set of column labels to be kept in the dataframe.
+                set values=1 to be sure that all variables are kept.
+                don't worry: essential columns like ID, coordinates
+                and time will always be kept.
+
+            scale: string; default 'maxabs'
+                specify how the data should be scaled.
+                possible values are None, maxabs, minmax, std.
+                for more input, check out: help(pcdl.scaler)
+
+        output:
+            annmcds: anndata object
+                for this one time step.
+
+        description:
+            function to transform a mcds time step into an anndata object
+            for downstream analysis.
+        """
+        # processing
+        if self.verbose:
+            print(f'processing: 1/1 {round(self.get_time(),9)}[min] mcds into anndata obj.')
+        df_cell = self.get_cell_df(values=values, drop=drop, keep=keep)
+        df_count, df_obs, d_obsm, d_obsp, d_uns = _anndextract(
+            df_cell = df_cell,
+            scale = scale,
+            graph_attached = self.get_attached_graph_dict(),
+            graph_neighbor = self.get_neighbor_graph_dict(),
+            graph_method = self.get_physicell_version(),
+        )
+        annmcds = ad.AnnData(
+            X = df_count,
+            obs = df_obs,
+            obsm = d_obsm,
+            obsp = d_obsp,
+            uns = d_uns
+        )
+        # output
+        return annmcds
+
+
+
     ## LOAD DATA  ##
 
     def _read_xml(self, xmlfile, output_path='.'):
         """
         input:
-            self: pyMCDS class instance.
+            self: TimeStep class instance.
 
             xmlfile: string
                 name of the xml file with or without path
@@ -2516,11 +2811,11 @@ class pyMCDS:
                 the PhysiCell output files are stored.
 
         output:
-            self: pyMCDS class instance with loaded data.
+            self: TimeStep class instance with loaded data.
 
         description:
             internal function to load the data from the PhysiCell output files
-            into the pyMCDS instance.
+            into the TimeStep instance.
         """
         #####################
         # path and filename #
@@ -2870,12 +3165,12 @@ class pyMCDS:
                 print(f'reading: {s_cellpathfile}')
         except ValueError:  # hack: some old PhysiCell versions generates a corrupt cells.mat file, if there are zero cells.
             if self.verbose:
-                print(f'Warning @ pyMCDS._read_xml : corrupt {s_cellpathfile} detected!\nassuming time step with zero cells because of a known bug in PhysiCell MultiCellDS version 0.5 output.')
+                print(f'Warning @ TimeStep._read_xml : corrupt {s_cellpathfile} detected!\nassuming time step with zero cells because of a known bug in PhysiCell MultiCellDS version 0.5 output.')
             ar_cell = np.empty([len(ls_variable),0])
 
         # check for column label mapping error (as good as it gets)
         if (ar_cell.shape[0] != len(ls_variable)):
-            sys.exit(f'Error @ pyMCDS._read_xml : extracted column label list leng {len(ls_variable)} and data array shape {ar_cell.shape} are incompatible!')
+            sys.exit(f'Error @ TimeStep._read_xml : extracted column label list leng {len(ls_variable)} and data array shape {ar_cell.shape} are incompatible!')
 
         # metadata cell_type label:id mapping detection ~ label information lost (bronze quality)
         if not b_celltype:
@@ -2964,7 +3259,7 @@ class pyMCDS:
                     df_physiboss[f'node_{s_node}'] = df_physiboss.state.str.find(s_node) > -1
 
             elif self.verbose:
-                print(f'Warning @ pyMCDS._read_xml : physiboss file missing {s_intracellpathfile}.')
+                print(f'Warning @ TimeStep._read_xml : physiboss file missing {s_intracellpathfile}.')
 
             else:
                 pass
