+++ test/test_timeseries_3d.py
@@ -7,7 +7,7 @@
 # license: BSD 3-Clause
 #
 # description:
-#   pytest unit test library for the pcdl library pyMCDSts class.
+#   pytest unit test library for the pcdl library TimeSeries class.
 #   + https://docs.pytest.org/
 #
 #   note:
@@ -44,60 +44,60 @@ if not os.path.exists(s_path_3d):
 ## data loading related functions ##
 
 class TestPyMcdsTs3DInit(object):
-    ''' tests for loading a pcdl.pyMCDSts data set. '''
+    ''' tests for loading a pcdl.TimeSeries data set. '''
 
     def test_mcdsts_set_verbose_true(self):
-        mcdsts = pcdl.pyMCDSts(s_path_3d, load=False, verbose=False)
+        mcdsts = pcdl.TimeSeries(s_path_3d, load=False, verbose=False)
         mcdsts.set_verbose_true()
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (mcdsts.verbose)
 
     def test_mcdsts_set_verbose_false(self):
-        mcdsts = pcdl.pyMCDSts(s_path_3d, load=False, verbose=True)
+        mcdsts = pcdl.TimeSeries(s_path_3d, load=False, verbose=True)
         mcdsts.set_verbose_false()
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (not mcdsts.verbose)
 
     ## get_xmlfile and read_mcds command and get_mcds_list ##
     def test_mcdsts_get_xmlfile_list(self):
-        mcdsts = pcdl.pyMCDSts(s_path_3d, load=False, verbose=True)
+        mcdsts = pcdl.TimeSeries(s_path_3d, load=False, verbose=True)
         ls_xmlfile = mcdsts.get_xmlfile_list()
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (ls_xmlfile[0] == 'output00000000.xml') and \
               (ls_xmlfile[-1] == 'output00000024.xml') and \
               (len(ls_xmlfile) == 25)
 
     def test_mcdsts_get_mcds_list(self):
-        mcdsts = pcdl.pyMCDSts(s_path_3d, load=True, verbose=True)
+        mcdsts = pcdl.TimeSeries(s_path_3d, load=True, verbose=True)
         l_mcds = mcdsts.get_mcds_list()
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
-              (str(type(mcdsts.l_mcds[0])) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
-              (str(type(mcdsts.l_mcds[-1])) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
+              (str(type(mcdsts.l_mcds[0])) == "<class 'pcdl.timestep.TimeStep'>") and \
+              (str(type(mcdsts.l_mcds[-1])) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcdsts.l_mcds[0].get_time() == 0) and \
               (mcdsts.l_mcds[-1].get_time() == 1440) and \
               (len(mcdsts.l_mcds) == 25) and \
               (mcdsts.l_mcds == l_mcds)
 
     def test_mcdsts_read_mcds(self):
-        mcdsts = pcdl.pyMCDSts(s_path_3d, load=False, verbose=True)
+        mcdsts = pcdl.TimeSeries(s_path_3d, load=False, verbose=True)
         l_mcds_loadfalse  = mcdsts.get_mcds_list()
         mcdsts.read_mcds()
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
-              (str(type(mcdsts.l_mcds[0])) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
-              (str(type(mcdsts.l_mcds[-1])) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
+              (str(type(mcdsts.l_mcds[0])) == "<class 'pcdl.timestep.TimeStep'>") and \
+              (str(type(mcdsts.l_mcds[-1])) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcdsts.l_mcds[0].get_time() == 0) and \
               (mcdsts.l_mcds[-1].get_time() == 1440) and \
               (len(mcdsts.l_mcds) == 25) and \
               (l_mcds_loadfalse is None)
 
     def test_mcdsts_read_mcds_xmlfilelist(self):
-        mcdsts = pcdl.pyMCDSts(s_path_3d, load=False, verbose=True)
+        mcdsts = pcdl.TimeSeries(s_path_3d, load=False, verbose=True)
         ls_xmlfile = mcdsts.get_xmlfile_list()
         ls_xmlfile = ls_xmlfile[-3:]
         l_mcds = mcdsts.read_mcds(ls_xmlfile)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
-              (str(type(mcdsts.l_mcds[0])) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
-              (str(type(mcdsts.l_mcds[-1])) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
+              (str(type(mcdsts.l_mcds[0])) == "<class 'pcdl.timestep.TimeStep'>") and \
+              (str(type(mcdsts.l_mcds[-1])) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcdsts.l_mcds[0].get_time() == 1320) and \
               (mcdsts.l_mcds[-1].get_time() == 1440) and \
               (len(ls_xmlfile) == 3) and \
@@ -108,12 +108,12 @@ class TestPyMcdsTs3DInit(object):
 ## micro environment related functions ##
 
 class TestPyMcdsTs3DMicroenv(object):
-    ''' tests for pcdl.pyMCDS micro environment related functions. '''
-    mcdsts = pcdl.pyMCDSts(s_path_3d, verbose=True)
+    ''' tests for pcdl.TimeStep micro environment related functions. '''
+    mcdsts = pcdl.TimeSeries(s_path_3d, verbose=True)
 
     def test_mcdsts_get_conc_df(self, mcdsts=mcdsts):
         ldf_conc = mcdsts.get_conc_df(values=2, drop=set(), keep=set(), collapse=False)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(ldf_conc)) == "<class 'list'>") and \
               (str(type(ldf_conc[0])) == "<class 'pandas.core.frame.DataFrame'>") and \
               (ldf_conc[0].shape == (1331, 9)) and \
@@ -122,13 +122,13 @@ class TestPyMcdsTs3DMicroenv(object):
 
     def test_mcdsts_get_conc_df_collapse(self, mcdsts=mcdsts):
         df_conc = mcdsts.get_conc_df(values=2, drop=set(), keep=set(), collapse=True)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(df_conc)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_conc.shape == (33275, 11))
 
     def test_mcdsts_get_conc_attribute(self, mcdsts=mcdsts):
         dl_conc = mcdsts.get_conc_attribute(values=1, drop=set(), keep=set(), allvalues=False)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(dl_conc)) == "<class 'dict'>") and \
               (str(type(dl_conc['oxygen'])) == "<class 'list'>") and \
               (str(type(dl_conc['oxygen'][0])) == "<class 'float'>") and \
@@ -137,7 +137,7 @@ class TestPyMcdsTs3DMicroenv(object):
 
     def test_mcdsts_get_conc_attribute_values(self, mcdsts=mcdsts):
         dl_conc = mcdsts.get_conc_attribute(values=2, drop=set(), keep=set(), allvalues=False)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(dl_conc)) == "<class 'dict'>") and \
               (str(type(dl_conc['oxygen'])) == "<class 'list'>") and \
               (str(type(dl_conc['oxygen'][0])) == "<class 'float'>") and \
@@ -146,7 +146,7 @@ class TestPyMcdsTs3DMicroenv(object):
 
     def test_mcdsts_get_conc_attribute_allvalues(self, mcdsts=mcdsts):
         dl_conc = mcdsts.get_conc_attribute(values=1, drop=set(), keep=set(), allvalues=True)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(dl_conc)) == "<class 'dict'>") and \
               (str(type(dl_conc['oxygen'])) == "<class 'list'>") and \
               (str(type(dl_conc['oxygen'][0])) == "<class 'float'>") and \
@@ -171,7 +171,7 @@ class TestPyMcdsTs3DMicroenv(object):
             ext = 'jpeg',
             figbgcolor = None,  # test if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (ls_pathfile[0].replace('\\','/').endswith('/pcdl/output_3d/conc_oxygen_z-5.0/output00000000_oxygen.jpeg')) and \
               (os.path.exists(ls_pathfile[0])) and \
               (os.path.getsize(ls_pathfile[0]) > 2**10) and \
@@ -187,19 +187,19 @@ class TestPyMcdsTs3DMicroenv(object):
             focus = 'oxygen',
             z_slice = 0.0,  # jump over if
             extrema = [0, 38],  # jump over if
-            #alpha = 1,  # pyMCDS
-            #fill = True,  # pyMCDS
-            #cmap = 'viridis',  # pyMCDS
+            #alpha = 1,  # TimeStep
+            #fill = True,  # TimeStep
+            #cmap = 'viridis',  # TimeStep
             title = 'abc',  # test non default
-            #grid = True,  # pyMCDS
+            #grid = True,  # TimeStep
             xlim = [-31, 301],  # jump over if
             ylim = [-21, 201],  # jump over if
-            #xyequal = True,  # pyMCDS
+            #xyequal = True,  # TimeStep
             figsizepx = [641, 481],  # test non even pixel
             ext = None,
             figbgcolor = 'yellow',  # jump over if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(l_fig[0])) == "<class 'matplotlib.figure.Figure'>") and \
               (str(type(l_fig[-1])) == "<class 'matplotlib.figure.Figure'>") and \
               (len(l_fig) == 25)
@@ -208,7 +208,7 @@ class TestPyMcdsTs3DMicroenv(object):
 
     def test_mcdsts_make_conc_vtk(self, mcdsts=mcdsts):
         ls_pathfile = mcdsts.make_conc_vtk(visualize=False)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (ls_pathfile[0].endswith('/pcdl/output_3d/output00000000_conc.vtr')) and \
               (ls_pathfile[-1].endswith('/pcdl/output_3d/output00000024_conc.vtr')) and \
               (os.path.exists(ls_pathfile[0])) and \
@@ -223,12 +223,12 @@ class TestPyMcdsTs3DMicroenv(object):
 ## cell related functions ##
 
 class TestPyMcds3DCell(object):
-    ''' tests for pcdl.pyMCDS cell related functions. '''
-    mcdsts = pcdl.pyMCDSts(s_path_3d, verbose=False)
+    ''' tests for pcdl.TimeStep cell related functions. '''
+    mcdsts = pcdl.TimeSeries(s_path_3d, verbose=False)
 
     def test_mcdsts_get_cell_df(self, mcdsts=mcdsts):
         ldf_cell = mcdsts.get_cell_df(values=2, drop=set(), keep=set(), collapse=False)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(ldf_cell)) == "<class 'list'>") and \
               (str(type(ldf_cell[0])) == "<class 'pandas.core.frame.DataFrame'>") and \
               (ldf_cell[0].shape[0] > 9) and \
@@ -239,14 +239,14 @@ class TestPyMcds3DCell(object):
 
     def test_mcdsts_get_cell_df_collapse(self, mcdsts=mcdsts):
         df_cell = mcdsts.get_cell_df(values=2, drop=set(), keep=set(), collapse=True)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 73)
 
     def test_mcdsts_get_cell_attribute(self, mcdsts=mcdsts):
         dl_cell = mcdsts.get_cell_attribute(values=1, drop=set(), keep=set(), allvalues=False)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(dl_cell)) == "<class 'dict'>") and \
               (str(type(dl_cell['dead'])) == "<class 'list'>") and \
               (str(type(dl_cell['dead'][0])) == "<class 'bool'>") and \
@@ -264,13 +264,13 @@ class TestPyMcds3DCell(object):
 
     def test_mcdsts_get_cell_attribute_values(self, mcdsts=mcdsts):
         dl_cell = mcdsts.get_cell_attribute(values=2, drop=set(), keep=set(), allvalues=False)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(dl_cell)) == "<class 'dict'>") and \
               (len(dl_cell.keys()) == 60)
 
     def test_mcdsts_get_cell_attribute_allvalues(self, mcdsts=mcdsts):
         dl_cell = mcdsts.get_cell_attribute(values=1, drop=set(), keep=set(), allvalues=True)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(dl_cell)) == "<class 'dict'>") and \
               (str(type(dl_cell['dead'])) == "<class 'list'>") and \
               (str(type(dl_cell['dead'][0])) == "<class 'bool'>") and \
@@ -299,13 +299,13 @@ class TestPyMcds3DCell(object):
             #legend_loc='lower left',  # matplotlib
             xlim = None,  # test if
             ylim = None,  # test if
-            #xyequal = True,  # pyMCDS
+            #xyequal = True,  # TimeStep
             s = None,  # test if
             figsizepx = None,  # case extract from initial.svg
             ext = 'jpeg',
             figbgcolor = None,  # test if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (ls_pathfile[0].replace('\\','/').endswith('/pcdl/output_3d/cell_pressure_z-5.0/output00000000_pressure.jpeg')) and \
               (os.path.exists(ls_pathfile[0])) and \
               (os.path.getsize(ls_pathfile[0]) > 2**10) and \
@@ -321,20 +321,20 @@ class TestPyMcds3DCell(object):
             focus='cell_type',  # case categorical
             z_slice = 0.0,   # jump over if
             z_axis = None,  # test iff  categorical
-            #alpha = 1,  # pyMCDS
-            #cmap = 'viridis',  # pyMCDS
+            #alpha = 1,  # TimeStep
+            #cmap = 'viridis',  # TimeStep
             title = 'abc',  # test non default
-            #grid = True,  # pyMCDS
-            #legend_loc='lower left',  # pyMCDS
+            #grid = True,  # TimeStep
+            #legend_loc='lower left',  # TimeStep
             xlim = None,  # test if
             ylim = None,  # test if
-            #xyequal = True,  # pyMCDS
+            #xyequal = True,  # TimeStep
             s = None,  # test if
             figsizepx = [641, 481],  # test case non even pixel number
             ext = None,
             figbgcolor = 'cyan',  # jump over if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(l_fig[0])) == "<class 'matplotlib.figure.Figure'>") and \
               (str(type(l_fig[-1])) == "<class 'matplotlib.figure.Figure'>") and \
               (len(l_fig) == 25)
@@ -342,7 +342,7 @@ class TestPyMcds3DCell(object):
 
     def test_mcdsts_make_cell_vtk(self, mcdsts=mcdsts):
         ls_pathfile = mcdsts.make_cell_vtk(visualize=False)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (ls_pathfile[0].endswith('/pcdl/output_3d/output00000000_cell.vtp')) and \
               (ls_pathfile[-1].endswith('/pcdl/output_3d/output00000024_cell.vtp')) and \
               (os.path.exists(ls_pathfile[0])) and \
@@ -356,13 +356,13 @@ class TestPyMcds3DCell(object):
 
 ## graph related functions ##
 class TestPyMcds3DGraph(object):
-    ''' tests for pcdl.pyMCDS graph related functions. '''
-    mcdsts = pcdl.pyMCDSts(s_path_3d, verbose=False)
+    ''' tests for pcdl.TimeStep graph related functions. '''
+    mcdsts = pcdl.TimeSeries(s_path_3d, verbose=False)
 
     ## graph related functions ##
     def test_mcdsts_get_graph_gml_attached_defaultattr(self, mcdsts=mcdsts):
         ls_pathfile = mcdsts.make_graph_gml(graph_type='attached', edge_attribute=True, node_attribute=[])
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (ls_pathfile[0].endswith('/pcdl/output_3d/output00000000_attached.gml')) and \
               (ls_pathfile[-1].endswith('/pcdl/output_3d/output00000024_attached.gml')) and \
               (os.path.exists(ls_pathfile[0])) and \
@@ -373,7 +373,7 @@ class TestPyMcds3DGraph(object):
 
     def test_mcdsts_get_graph_gml_neighbor_noneattr(self, mcdsts=mcdsts):
         ls_pathfile = mcdsts.make_graph_gml(graph_type='neighbor', edge_attribute=False, node_attribute=[])
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (ls_pathfile[0].endswith('/pcdl/output_3d/output00000000_neighbor.gml')) and \
               (ls_pathfile[-1].endswith('/pcdl/output_3d/output00000024_neighbor.gml')) and \
               (os.path.exists(ls_pathfile[0])) and \
@@ -384,7 +384,7 @@ class TestPyMcds3DGraph(object):
 
     def test_mcdsts_get_graph_gml_neighbor_allattr(self, mcdsts=mcdsts):
         ls_pathfile = mcdsts.make_graph_gml(graph_type='neighbor', edge_attribute=True, node_attribute=['dead','cell_count_voxel','cell_density_micron3','cell_type'])
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (ls_pathfile[0].endswith('/pcdl/output_3d/output00000000_neighbor.gml')) and \
               (ls_pathfile[-1].endswith('/pcdl/output_3d/output00000024_neighbor.gml')) and \
               (os.path.exists(ls_pathfile[0])) and \
@@ -395,13 +395,13 @@ class TestPyMcds3DGraph(object):
 
 ## graph related functions ##
 class TestPyMcds3DGraph(object):
-    ''' tests for pcdl.pyMCDS graph related functions. '''
-    mcdsts = pcdl.pyMCDSts(s_path_3d, verbose=False)
+    ''' tests for pcdl.TimeStep graph related functions. '''
+    mcdsts = pcdl.TimeSeries(s_path_3d, verbose=False)
 
     ## graph related functions ##
     def test_mcdsts_make_ome_tiff_defaultattr_00(self, mcdsts=mcdsts):
         la_ometiff = mcdsts.make_ome_tiff(cell_attribute='ID', conc_cutoff={}, focus=None, file=False, collapse=False)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (type(la_ometiff) is list) and \
               (type(la_ometiff[0]) is np.ndarray) and \
               (type(la_ometiff[-1]) is np.ndarray) and \
@@ -413,7 +413,7 @@ class TestPyMcds3DGraph(object):
 
     def test_mcdsts_make_ome_tiff_defaultattr_01(self, mcdsts=mcdsts):
         a_ometiff = mcdsts.make_ome_tiff(cell_attribute='ID', conc_cutoff={}, focus=None, file=False, collapse=True)
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (type(a_ometiff) is np.ndarray) and \
               (a_ometiff.dtype == np.float32) and \
               (a_ometiff.shape == (25, 4, 11, 200, 300))
@@ -422,8 +422,8 @@ class TestPyMcds3DGraph(object):
 ## timeseries related functions ##
 
 class TestPyMcds3DTimeseries(object):
-    ''' tests for pcdl.pyMCDS graph related functions. '''
-    mcdsts = pcdl.pyMCDSts(s_path_3d, verbose=False)
+    ''' tests for pcdl.TimeStep graph related functions. '''
+    mcdsts = pcdl.TimeSeries(s_path_3d, verbose=False)
 
     ## plot_timeseries command ##
     def test_mcdsts_plot_timeseries_none_none_none_cell_ax_jpeg(self, mcdsts=mcdsts):
@@ -453,7 +453,7 @@ class TestPyMcds3DTimeseries(object):
             ext = 'jpeg',  # test if else {'jpeg', None}
             figbgcolor = None  # test if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (s_pathfile.endswith('/pcdl/output_3d/timeseries_cell_total_count.jpeg')) and \
               (os.path.exists(s_pathfile))
         os.remove(s_pathfile)
@@ -484,7 +484,7 @@ class TestPyMcds3DTimeseries(object):
             ext = None,  # test if else {'jpeg', None}
             figbgcolor = None  # test if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
 
@@ -514,7 +514,7 @@ class TestPyMcds3DTimeseries(object):
             ext = None,  # test if else {'jpeg', None}
             figbgcolor = None  # test if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
 
@@ -544,7 +544,7 @@ class TestPyMcds3DTimeseries(object):
             ext = None,  # test if else {'jpeg', None}
             figbgcolor = None  # test if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
 
@@ -575,7 +575,7 @@ class TestPyMcds3DTimeseries(object):
             ext = 'jpeg',  # test if else {'jpeg', None}
             figbgcolor = None  # test if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (s_pathfile.endswith('/pcdl/output_3d/timeseries_conc_total_count.jpeg')) and \
               (os.path.exists(s_pathfile))
         os.remove(s_pathfile)
@@ -607,7 +607,7 @@ class TestPyMcds3DTimeseries(object):
             ext = None,  # test if else {'jpeg', None}
             figbgcolor = None  # test if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
 
@@ -637,7 +637,7 @@ class TestPyMcds3DTimeseries(object):
             ext = None,  # test if else {'jpeg', None}
             figbgcolor = None  # test if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
 
@@ -667,7 +667,7 @@ class TestPyMcds3DTimeseries(object):
             ext = None,  # test if else {'jpeg', None}
             figbgcolor = None  # test if
         )
-        assert(str(type(mcdsts)) == "<class 'pcdl.pyMCDSts.pyMCDSts'>") and \
+        assert(str(type(mcdsts)) == "<class 'pcdl.timeseries.TimeSeries'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
