+++ test/test_timestep_2d.py
@@ -7,7 +7,7 @@
 # license: BSD 3-Clause
 #
 # description:
-#   pytest unit test library for the pcdl library pyMCDS class.
+#   pytest unit test library for the pcdl library TimeStep class.
 #   + https://docs.pytest.org/
 #
 #   note:
@@ -40,28 +40,28 @@ if not os.path.exists(s_path_2d):
 ## data loading related functions ##
 
 class TestPyMcdsInit(object):
-    ''' tests for loading a pcdl.pyMCDS data set. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for loading a pcdl.TimeStep data set. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
     df_cell = mcds.get_cell_df()
     def test_mcds_init_microenv(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 122)
 
     def test_mcds_init_graph(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['attached_cells'])) == "<class 'dict'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['neighbor_cells'])) == "<class 'dict'>") and \
               (len(mcds.data['discrete_cells']['graph']['attached_cells']) > 9) and \
               (len(mcds.data['discrete_cells']['graph']['neighbor_cells']) > 9)
 
     def test_mcds_init_physiboss(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcds.data['discrete_cells']['physiboss'] == None)
 
     def test_mcds_init_settingxml(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (set(df_cell.columns).issuperset({'default_fusion_rates'})) and \
               (df_cell.shape[0] > 9) and \
@@ -69,28 +69,28 @@ class TestPyMcdsInit(object):
 
 
 class TestPyMcdsInitMicroenvFalse(object):
-    ''' tests for loading a pcdl.pyMCDS data set with microenv false. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=False, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for loading a pcdl.TimeStep data set with microenv false. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=False, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
     df_cell = mcds.get_cell_df()
     def test_mcds_init_microenv(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 116)
 
     def test_mcds_init_graph(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['attached_cells'])) == "<class 'dict'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['neighbor_cells'])) == "<class 'dict'>") and \
               (len(mcds.data['discrete_cells']['graph']['attached_cells']) > 9) and \
               (len(mcds.data['discrete_cells']['graph']['neighbor_cells']) > 9)
 
     def test_mcds_init_physiboss(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcds.data['discrete_cells']['physiboss'] == None)
 
     def test_mcds_init_settingxml(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (set(df_cell.columns).issuperset({'default_fusion_rates'})) and \
               (df_cell.shape[0] > 9) and \
@@ -98,28 +98,28 @@ class TestPyMcdsInitMicroenvFalse(object):
 
 
 class TestPyMcdsInitGraphFalse(object):
-    ''' tests for loading a pcdl.pyMCDS data set with graph false. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=False, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for loading a pcdl.TimeStep data set with graph false. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=False, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
     df_cell = mcds.get_cell_df()
     def test_mcds_init_microenv(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 122)
 
     def test_mcds_init_graph(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['attached_cells'])) == "<class 'dict'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['neighbor_cells'])) == "<class 'dict'>") and \
               (len(mcds.data['discrete_cells']['graph']['attached_cells']) == 0) and \
               (len(mcds.data['discrete_cells']['graph']['neighbor_cells']) == 0)
 
     def test_mcds_init_physiboss(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcds.data['discrete_cells']['physiboss'] == None)
 
     def test_mcds_init_settingxml(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (set(df_cell.columns).issuperset({'default_fusion_rates'})) and \
               (df_cell.shape[0] > 9) and \
@@ -127,28 +127,28 @@ class TestPyMcdsInitGraphFalse(object):
 
 
 class TestPyMcdsInitPhysibossFalse(object):
-    ''' tests for loading a pcdl.pyMCDS data set with physiboss false. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=False, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for loading a pcdl.TimeStep data set with physiboss false. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=False, settingxml='PhysiCell_settings.xml', verbose=True)
     df_cell = mcds.get_cell_df()
     def test_mcds_init_microenv(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 122)
 
     def test_mcds_init_graph(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['attached_cells'])) == "<class 'dict'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['neighbor_cells'])) == "<class 'dict'>") and \
               (len(mcds.data['discrete_cells']['graph']['attached_cells']) > 9) and \
               (len(mcds.data['discrete_cells']['graph']['neighbor_cells']) > 9)
 
     def test_mcds_init_physiboss(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcds.data['discrete_cells']['physiboss'] == None)
 
     def test_mcds_init_settingxml(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (set(df_cell.columns).issuperset({'default_fusion_rates'})) and \
               (df_cell.shape[0] > 9) and \
@@ -156,28 +156,28 @@ class TestPyMcdsInitPhysibossFalse(object):
 
 
 class TestPyMcdsInitSettingxmlFalse(object):
-    ''' tests for loading a pcdl.pyMCDS data set with settingxml false. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml=False, verbose=True)
+    ''' tests for loading a pcdl.TimeStep data set with settingxml false. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml=False, verbose=True)
     df_cell = mcds.get_cell_df()
     def test_mcds_init_microenv(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 122)
 
     def test_mcds_init_graph(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['attached_cells'])) == "<class 'dict'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['neighbor_cells'])) == "<class 'dict'>") and \
               (len(mcds.data['discrete_cells']['graph']['attached_cells']) > 9) and \
               (len(mcds.data['discrete_cells']['graph']['neighbor_cells']) > 9)
 
     def test_mcds_init_physiboss(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcds.data['discrete_cells']['physiboss'] == None)
 
     def test_mcds_init_settingxml(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (set(df_cell.columns).issuperset({'default_fusion_rates'})) and \
               (df_cell.shape[0] > 9) and \
@@ -185,28 +185,28 @@ class TestPyMcdsInitSettingxmlFalse(object):
 
 
 class TestPyMcdsInitSettingxmlNone(object):
-    ''' tests for loading a pcdl.pyMCDS data set with settingxml none. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml=None, verbose=True)
+    ''' tests for loading a pcdl.TimeStep data set with settingxml none. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml=None, verbose=True)
     df_cell = mcds.get_cell_df()
     def test_mcds_init_microenv(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 122)
 
     def test_mcds_init_graph(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['attached_cells'])) == "<class 'dict'>") and \
               (str(type(mcds.data['discrete_cells']['graph']['neighbor_cells'])) == "<class 'dict'>") and \
               (len(mcds.data['discrete_cells']['graph']['attached_cells']) > 9) and \
               (len(mcds.data['discrete_cells']['graph']['neighbor_cells']) > 9)
 
     def test_mcds_init_physiboss(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcds.data['discrete_cells']['physiboss'] == None)
 
     def test_mcds_init_settingxml(self, mcds=mcds, df_cell=df_cell):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (set(df_cell.columns).issuperset({'default_fusion_rates'})) and \
               (df_cell.shape[0] > 9) and \
@@ -214,66 +214,66 @@ class TestPyMcdsInitSettingxmlNone(object):
 
 
 class TestPyMcdsInitVerboseTrue(object):
-    ''' tests for loading a pcdl.pyMCDS data set and set_verbose_false function. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for loading a pcdl.TimeStep data set and set_verbose_false function. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
 
     def test_mcds_verbose_true(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcds.verbose)
 
     def test_mcds_set_verbose_false(self, mcds=mcds):
         mcds.set_verbose_false()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (not mcds.verbose)
 
 
 class TestPyMcdsInitVerboseFalse(object):
-    ''' tests for loading a pcdl.pyMCDS data set and set_verbose_true function. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=False)
+    ''' tests for loading a pcdl.TimeStep data set and set_verbose_true function. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=False)
 
     def test_mcds_verbose_false(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (not mcds.verbose)
 
     def test_mcds_set_verbose_true(self, mcds=mcds):
         mcds.set_verbose_true()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcds.verbose)
 
 
 ## metadata related functions ##
 
 class TestPyMcdsMetadata(object):
-    ''' tests for pcdl.pyMCDS metadata related functions. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for pcdl.TimeStep metadata related functions. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
 
     def test_mcds_get_multicellds_version(self, mcds=mcds):
         s_mcdsversion = mcds.get_multicellds_version()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(s_mcdsversion)) == "<class 'str'>") and \
               (s_mcdsversion == 'MultiCellDS_2')
 
     def test_mcds_get_physicell_version(self, mcds=mcds):
         s_pcversion = mcds.get_physicell_version()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(s_pcversion)) == "<class 'str'>") and \
               (s_pcversion == 'PhysiCell_1.14.1')
 
     def test_mcds_get_timestamp(self, mcds=mcds):
         s_timestamp = mcds.get_timestamp()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(s_timestamp)) == "<class 'str'>") and \
               (s_timestamp == '2025-01-05T08:08:22Z')
 
     def test_mcds_get_time(self, mcds=mcds):
         r_time = mcds.get_time()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(r_time)) == "<class 'float'>") and \
               (r_time == 1440.0)
 
     def test_mcds_get_runtime(self, mcds=mcds):
         r_runtime = mcds.get_runtime()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(r_runtime)) == "<class 'float'>") and \
               (r_runtime == 1.952156)
 
@@ -281,12 +281,12 @@ class TestPyMcdsMetadata(object):
 ## setting related functions ##
 
 class TestPyMcdsSetting(object):
-    ''' tests for pcdl.pyMCDS setting related functions. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for pcdl.TimeStep setting related functions. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
 
     def test_mcds_get_unit_dict(self, mcds=mcds):
         ds_unit = mcds.get_unit_dict()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ds_unit)) == "<class 'dict'>") and \
               (len(ds_unit) == 108) and \
               (ds_unit['oxygen'] == 'dimensionless')
@@ -295,12 +295,12 @@ class TestPyMcdsSetting(object):
 ## mesh related functions ##
 
 class TestPyMcdsMesh(object):
-    ''' tests for pcdl.pyMCDS mesh related functions. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for pcdl.TimeStep mesh related functions. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
 
     def test_mcds_get_voxel_ijk_range(self, mcds=mcds):
         lti_range = mcds.get_voxel_ijk_range()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(lti_range)) == "<class 'list'>") and \
               (str(type(lti_range[0])) == "<class 'tuple'>") and \
               (str(type(lti_range[0][0])) == "<class 'int'>") and \
@@ -308,7 +308,7 @@ class TestPyMcdsMesh(object):
 
     def test_mcds_get_mesh_mnp_range(self, mcds=mcds):
         ltr_range = mcds.get_mesh_mnp_range()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ltr_range)) == "<class 'list'>") and \
               (str(type(ltr_range[0])) == "<class 'tuple'>") and \
               (str(type(ltr_range[0][0])) == "<class 'numpy.float64'>") and \
@@ -316,7 +316,7 @@ class TestPyMcdsMesh(object):
 
     def test_mcds_get_xyz_range(self, mcds=mcds):
         ltr_range = mcds.get_xyz_range()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ltr_range)) == "<class 'list'>") and \
               (str(type(ltr_range[0])) == "<class 'tuple'>") and \
               (str(type(ltr_range[0][0])) == "<class 'numpy.float64'>") and \
@@ -324,7 +324,7 @@ class TestPyMcdsMesh(object):
 
     def test_mcds_get_voxel_ijk_axis(self, mcds=mcds):
         lai_axis = mcds.get_voxel_ijk_axis()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(lai_axis)) == "<class 'list'>") and \
               (str(type(lai_axis[0])) == "<class 'numpy.ndarray'>") and \
               (str(type(lai_axis[0][0])).startswith("<class 'numpy.int")) and \
@@ -335,7 +335,7 @@ class TestPyMcdsMesh(object):
 
     def test_mcds_get_mesh_mnp_axis(self, mcds=mcds):
         lar_axis = mcds.get_mesh_mnp_axis()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(lar_axis)) == "<class 'list'>") and \
               (str(type(lar_axis[0])) == "<class 'numpy.ndarray'>") and \
               (str(type(lar_axis[0][0])) == "<class 'numpy.float64'>") and \
@@ -346,14 +346,14 @@ class TestPyMcdsMesh(object):
 
     def test_mcds_get_mesh_flat_false(self, mcds=mcds):
         aar_mesh = mcds.get_mesh(flat=False)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(aar_mesh)) == "<class 'numpy.ndarray'>") and \
               (aar_mesh.dtype == np.float64) and \
               (aar_mesh.shape == (3, 11, 11, 1))
 
     def test_mcds_get_mesh_flat_true(self, mcds=mcds):
         aar_mesh = mcds.get_mesh(flat=True)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(aar_mesh)) == "<class 'numpy.ndarray'>") and \
               (aar_mesh.dtype == np.float64) and \
               (aar_mesh.shape == (2, 11, 11))
@@ -361,7 +361,7 @@ class TestPyMcdsMesh(object):
     def test_mcds_get_mesh_2d(self, mcds=mcds):
         aar_mesh_flat = mcds.get_mesh(flat=True)
         aar_mesh_2d = mcds.get_mesh_2D()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(aar_mesh_2d)) == "<class 'numpy.ndarray'>") and \
               (aar_mesh_2d.dtype == np.float64) and \
               (aar_mesh_2d.shape == (2, 11, 11))
@@ -374,7 +374,7 @@ class TestPyMcdsMesh(object):
         er_p_cube = set(ar_p_cube.flatten())
         # linear coordinates
         aar_voxel = mcds.get_mesh_coordinate()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(aar_voxel)) == "<class 'numpy.ndarray'>") and \
               (aar_voxel.dtype == np.float64) and \
               (aar_voxel.shape == (3, 121)) and \
@@ -384,14 +384,14 @@ class TestPyMcdsMesh(object):
 
     def test_mcds_get_voxel_volume(self, mcds=mcds):
         r_volume = mcds.get_voxel_volume()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(r_volume)) == "<class 'numpy.float64'>") and \
               (r_volume == 6000.0)
 
     # bue: check else in 3D
     def test_mcds_get_mesh_spacing(self, mcds=mcds):
         lr_spacing = mcds.get_mesh_spacing()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(lr_spacing)) == "<class 'list'>") and \
               (str(type(lr_spacing[0])) == "<class 'numpy.float64'>") and \
               (str(type(lr_spacing[1])) == "<class 'numpy.float64'>") and \
@@ -400,14 +400,14 @@ class TestPyMcdsMesh(object):
 
     def test_mcds_get_voxel_spacing(self, mcds=mcds):
         lr_spacing = mcds.get_voxel_spacing()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(lr_spacing)) == "<class 'list'>") and \
               (str(type(lr_spacing[0])) == "<class 'numpy.float64'>") and \
               (str(type(lr_spacing[-1])) == "<class 'numpy.float64'>") and \
               (lr_spacing == [30.0, 20.0, 10.0])
 
     def test_mcds_is_in_mesh(self, mcds=mcds):
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (mcds.is_in_mesh(x=0, y=0, z=0, halt=False)) and \
               (not mcds.is_in_mesh(x=301, y=0, z=0, halt=False)) and \
               (not mcds.is_in_mesh(x=0, y=201, z=0, halt=False)) and \
@@ -418,7 +418,7 @@ class TestPyMcdsMesh(object):
         li_mesh_1 = mcds.get_mesh_mnp(x=15, y=10, z=0, is_in_mesh=True) # if b_calc
         li_mesh_2 = mcds.get_mesh_mnp(x=30, y=20, z=0, is_in_mesh=True) # if b_calc
         li_mesh_none = mcds.get_mesh_mnp(x=-31, y=-21, z=-6, is_in_mesh=True) # else b_calc
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(li_mesh_0)) == "<class 'list'>") and \
               (str(type(li_mesh_0[0])) == "<class 'numpy.float64'>") and \
               (li_mesh_0 == [-15.0, -10.0, 0.0]) and \
@@ -431,7 +431,7 @@ class TestPyMcdsMesh(object):
         li_voxel_1 = mcds.get_voxel_ijk(x=15, y=10, z=0, is_in_mesh=True) # if b_calc
         li_voxel_2 = mcds.get_voxel_ijk(x=30, y=20, z=0, is_in_mesh=True) # if b_calc
         li_voxel_none = mcds.get_voxel_ijk(x=-31, y=-21, z=-6, is_in_mesh=True) # else b_calc
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(li_voxel_0)) == "<class 'list'>") and \
               (str(type(li_voxel_0[0])) == "<class 'int'>") and \
               (li_voxel_0 == [0, 0, 0]) and \
@@ -443,95 +443,95 @@ class TestPyMcdsMesh(object):
 ## micro environment related functions ##
 
 class TestPyMcdsMicroenv(object):
-    ''' tests for pcdl.pyMCDS micro environment related functions. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for pcdl.TimeStep micro environment related functions. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
 
     def test_mcds_get_substrate_name(self, mcds=mcds):
         ls_substrate = mcds.get_substrate_list()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ls_substrate)) == "<class 'list'>") and \
               (str(type(ls_substrate[0])) == "<class 'str'>") and \
               (ls_substrate == ['oxygen','water'])
 
     def test_mcds_get_substrate_dict(self, mcds=mcds):
         ds_substrate = mcds.get_substrate_dict()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ds_substrate)) == "<class 'dict'>") and \
               (str(type(ds_substrate['0'])) == "<class 'str'>") and \
               (len(ds_substrate) == 2)
 
     def test_mcds_get_substrate_df(self, mcds=mcds):
         df_substrate = mcds.get_substrate_df()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_substrate)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_substrate.shape == (2, 2))
 
     def test_mcds_get_concentration_zslice_none(self, mcds=mcds):
         ar_conc = mcds.get_concentration(substrate='oxygen', z_slice=None)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ar_conc)) == "<class 'numpy.ndarray'>") and \
               (ar_conc.dtype == np.float64) and \
               (ar_conc.shape == (11, 11, 1))
 
     def test_mcds_get_concentration_zslice_meshcenter(self, mcds=mcds):
         ar_conc = mcds.get_concentration(substrate='oxygen', z_slice=0, halt=False)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ar_conc)) == "<class 'numpy.ndarray'>") and \
               (ar_conc.dtype == np.float64) and \
               (ar_conc.shape == (11, 11))
 
     def test_mcds_get_concentration_zslice_notmeshcenter(self, mcds=mcds):
         ar_conc = mcds.get_concentration(substrate='oxygen', z_slice=-3.333, halt=False)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ar_conc)) == "<class 'numpy.ndarray'>") and \
               (ar_conc.dtype == np.float64) and \
               (ar_conc.shape == (11, 11))
 
     def test_mcds_get_concentration_at_inmeash(self, mcds=mcds):
         ar_conc = mcds.get_concentration_at(x=0, y=0, z=0)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ar_conc)) == "<class 'numpy.ndarray'>") and \
               (ar_conc.dtype == np.float64) and \
               (ar_conc.shape == (2,))
 
     def test_mcds_get_concentration_at_notinmeash(self, mcds=mcds):
         ar_conc = mcds.get_concentration_at(x=-31, y=-21, z=-6)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (ar_conc is None)
 
     def test_mcds_get_conc_df(self, mcds=mcds):
         df_conc = mcds.get_conc_df(z_slice=None, halt=False, values=1, drop=set(), keep=set())
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_conc)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_conc.shape == (121, 11))
 
     def test_mcds_get_conc_df_zslice_center(self, mcds=mcds):
         df_conc = mcds.get_conc_df(z_slice=0, halt=False, values=1, drop=set(), keep=set())
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_conc)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_conc.shape == (121, 11))
 
     def test_mcds_get_conc_df_zslice_outofcenter(self, mcds=mcds):
         df_conc = mcds.get_conc_df(z_slice=-6, halt=False, values=1, drop=set(), keep=set())
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_conc)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_conc.shape == (121, 11))
 
     def test_mcds_get_conc_df_values(self, mcds=mcds):
         df_conc = mcds.get_conc_df(z_slice=None, halt=False, values=2, drop=set(), keep=set())
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_conc)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_conc.shape == (121, 11))
 
     def test_mcds_get_conc_df_drop(self, mcds=mcds):
         df_conc = mcds.get_conc_df(z_slice=None, halt=False, values=1, drop={'oxygen'}, keep=set())
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_conc)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_conc.shape == (121, 10))
 
     def test_mcds_get_conc_df_keep(self, mcds=mcds):
         df_conc = mcds.get_conc_df(z_slice=None, halt=False, values=1, drop=set(), keep={'oxygen'})
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_conc)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_conc.shape == (121, 10))
 
@@ -554,7 +554,7 @@ class TestPyMcdsMicroenv(object):
             ext = None, # test fig case
             figbgcolor = None,  # not at file
         )
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
 
@@ -578,7 +578,7 @@ class TestPyMcdsMicroenv(object):
             ext = 'tiff',  # test file case
             figbgcolor = 'yellow',  # jump over if
         )
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('/pcdl/output_2d/conc_oxygen_z0/output00000024_oxygen.tiff')) and \
               (os.path.exists(s_pathfile)) and \
               (os.path.getsize(s_pathfile) > 2**10)
@@ -586,7 +586,7 @@ class TestPyMcdsMicroenv(object):
 
     def test_mcds_make_conc_vtk(self, mcds=mcds):
         s_pathfile = mcds.make_conc_vtk(visualize=False)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('/pcdl/output_2d/output00000024_conc.vtr')) and \
               (os.path.exists(s_pathfile)) and \
               (os.path.getsize(s_pathfile) > 2**10)
@@ -595,61 +595,61 @@ class TestPyMcdsMicroenv(object):
 ## cell related functions ##
 
 class TestPyMcdsCell(object):
-    ''' tests for pcdl.pyMCDS cell related functions. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for pcdl.TimeStep cell related functions. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
 
     def test_mcds_get_celltype_list(self, mcds=mcds):
         ls_celltype = mcds.get_celltype_list()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ls_celltype)) == "<class 'list'>") and \
               (str(type(ls_celltype[0])) == "<class 'str'>") and \
               (len(ls_celltype) == 2)
 
     def test_mcds_get_celltype_dict(self, mcds=mcds):
         ds_celltype = mcds.get_celltype_dict()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(ds_celltype)) == "<class 'dict'>") and \
               (str(type(ds_celltype['0'])) == "<class 'str'>") and \
               (len(ds_celltype) == 2)
 
     def test_mcds_get_cell_df(self, mcds=mcds):
         df_cell = mcds.get_cell_df(values=1, drop=set(), keep=set())
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 122)
 
     def test_mcds_get_cell_df_values(self, mcds=mcds):
         df_cell = mcds.get_cell_df(values=2, drop=set(), keep=set())
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 66)
 
     def test_mcds_get_cell_df_drop(self, mcds=mcds):
         df_cell = mcds.get_cell_df(values=1, drop={'oxygen'}, keep=set())
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 121)
 
     def test_mcds_get_cell_df_keep(self, mcds=mcds):
         df_cell = mcds.get_cell_df(values=1, drop=set(), keep={'oxygen'})
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 9) and \
               (df_cell.shape[1] == 13)
 
     def test_mcds_get_cell_df_at_inmeash(self, mcds=mcds):
         df_cell = mcds.get_cell_df_at(x=0, y=0, z=0, values=1, drop=set(), keep=set())
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(df_cell)) == "<class 'pandas.core.frame.DataFrame'>") and \
               (df_cell.shape[0] > 0) and \
               (df_cell.shape[1] == 122)
 
     def test_mcds_get_cell_df_at_notinmeash(self, mcds=mcds):
         df_cell = mcds.get_cell_df_at(x=-31, y=-21, z=-6, values=1, drop=set(), keep=set())
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (df_cell is None)
 
     # scatter categorical
@@ -672,7 +672,7 @@ class TestPyMcdsCell(object):
             ext = None,  # test fig case
             figbgcolor = None,  # not a file
         )
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
 
@@ -695,7 +695,7 @@ class TestPyMcdsCell(object):
             ext = 'tiff',  # test file case
             figbgcolor = 'cyan',  # jump over if
         )
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('/pcdl/output_2d/cell_cell_type_z0/output00000024_cell_type.tiff')) and \
               (os.path.exists(s_pathfile)) and \
               (os.path.getsize(s_pathfile) > 2**10)
@@ -721,7 +721,7 @@ class TestPyMcdsCell(object):
             #ext = None,  # test fig case
             #figbgcolor = None,  # not a file
         )
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
 
@@ -745,7 +745,7 @@ class TestPyMcdsCell(object):
             #ext = None,  # test fig case
             #figbgcolor = None,  # not a file
         )
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
 
@@ -768,7 +768,7 @@ class TestPyMcdsCell(object):
             #ext = None,  # test fig case
             #figbgcolor = None,  # not a file
         )
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(fig)) == "<class 'matplotlib.figure.Figure'>")
         plt.close()
 
@@ -777,7 +777,7 @@ class TestPyMcdsCell(object):
             #attribute=['cell_type'],
             visualize=False,
         )
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('/pcdl/output_2d/output00000024_cell.vtp')) and \
               (os.path.exists(s_pathfile)) and \
               (os.path.getsize(s_pathfile) > 2**10)
@@ -785,7 +785,7 @@ class TestPyMcdsCell(object):
 
     def test_mcds_make_cell_vtk_attribute_zero(self, mcds=mcds):
         s_pathfile = mcds.make_cell_vtk(attribute=[], visualize=False)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('/pcdl/output_2d/output00000024_cell.vtp')) and \
               (os.path.exists(s_pathfile)) and \
               (os.path.getsize(s_pathfile) > 2**10)
@@ -796,7 +796,7 @@ class TestPyMcdsCell(object):
             attribute=['dead', 'cell_count_voxel', 'pressure', 'cell_type'],
             visualize=False,
         )
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('/pcdl/output_2d/output00000024_cell.vtp')) and \
               (os.path.exists(s_pathfile)) and \
               (os.path.getsize(s_pathfile) > 2**10)
@@ -806,14 +806,14 @@ class TestPyMcdsCell(object):
 ## graph related functions ##
 
 class TestPyMcdsGraph(object):
-    ''' tests for pcdl.pyMCDS graph related functions. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for pcdl.TimeStep graph related functions. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
 
     # graph dictionatry
     def test_mcds_get_attached_graph_dict(self, mcds=mcds):
         dei_graph = mcds.data['discrete_cells']['graph']['attached_cells']
         #print('graph attached:', sorted(dei_graph.items()))
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(dei_graph)) == "<class 'dict'>") and \
               (str(type(dei_graph[10])) == "<class 'set'>") and \
               (len(dei_graph[10]) == 0) and \
@@ -822,7 +822,7 @@ class TestPyMcdsGraph(object):
     def test_mcds_get_neighbor_graph_dict(self, mcds=mcds):
         dei_graph = mcds.data['discrete_cells']['graph']['neighbor_cells']
         #print('graph neighbor:', sorted(dei_graph.items()))
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(dei_graph)) == "<class 'dict'>") and \
               (str(type(dei_graph[10])) == "<class 'set'>") and \
               (len(dei_graph[10]) == 6) and \
@@ -835,7 +835,7 @@ class TestPyMcdsGraph(object):
         f = open(s_pathfile)
         s_file = f.read()
         f.close()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('pcdl/output_2d/output00000024_attached.gml')) and \
               (os.path.exists(s_pathfile)) and \
               (s_file.find('Creator "pcdl_v') > -1) and \
@@ -850,7 +850,7 @@ class TestPyMcdsGraph(object):
         f = open(s_pathfile)
         s_file = f.read()
         f.close()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('pcdl/output_2d/output00000024_attached.gml')) and \
               (os.path.exists(s_pathfile)) and \
               (s_file.find('Creator "pcdl_v') > -1) and \
@@ -865,7 +865,7 @@ class TestPyMcdsGraph(object):
         f = open(s_pathfile)
         s_file = f.read()
         f.close()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('pcdl/output_2d/output00000024_neighbor.gml')) and \
               (os.path.exists(s_pathfile)) and \
               (s_file.find('Creator "pcdl_v') > -1) and \
@@ -885,7 +885,7 @@ class TestPyMcdsGraph(object):
         f = open(s_pathfile)
         s_file = f.read()
         f.close()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('pcdl/output_2d/output00000024_neighbor.gml')) and \
               (os.path.exists(s_pathfile)) and \
               (s_file.find('Creator "pcdl_v') > -1) and \
@@ -900,7 +900,7 @@ class TestPyMcdsGraph(object):
         f = open(s_pathfile)
         s_file = f.read()
         f.close()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('pcdl/output_2d/output00000024_neighbor.gml')) and \
               (os.path.exists(s_pathfile)) and \
               (s_file.find('Creator "pcdl_v') > -1) and \
@@ -915,7 +915,7 @@ class TestPyMcdsGraph(object):
         f = open(s_pathfile)
         s_file = f.read()
         f.close()
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('pcdl/output_2d/output00000024_neighbor.gml')) and \
               (os.path.exists(s_pathfile)) and \
               (s_file.find('Creator "pcdl_v') > -1) and \
@@ -933,13 +933,13 @@ class TestPyMcdsGraph(object):
 ## ome tiff related functions ##
 
 class TestPyMcdsOmeTiff(object):
-    ''' tests for pcdl.pyMCDS graph related functions. '''
-    mcds = pcdl.pyMCDS(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
+    ''' tests for pcdl.TimeStep graph related functions. '''
+    mcds = pcdl.TimeStep(xmlfile=s_file_2d, output_path=s_path_2d, custom_data_type={}, microenv=True, graph=True, physiboss=True, settingxml='PhysiCell_settings.xml', verbose=True)
 
     ## ome tiff related functions ##
     def test_mcds_make_ome_tiff_default(self, mcds=mcds):
         s_pathfile = mcds.make_ome_tiff(cell_attribute='ID', conc_cutoff={}, focus=None, file=True)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (s_pathfile.replace('\\','/').endswith('pcdl/output_2d/output00000024_oxygen_water_default_blood_cells_ID.ome.tiff')) and \
               (os.path.exists(s_pathfile)) and \
               (os.path.getsize(s_pathfile) > 2**10)
@@ -947,7 +947,7 @@ class TestPyMcdsOmeTiff(object):
 
     def test_mcds_make_ome_tiff_bool(self, mcds=mcds):
         a_ometiff = mcds.make_ome_tiff(cell_attribute='dead', conc_cutoff={}, focus=None, file=False)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(a_ometiff)) == "<class 'numpy.ndarray'>") and \
               (a_ometiff.dtype == np.float32) and \
               (a_ometiff.shape == (4, 1, 200, 300)) and \
@@ -960,7 +960,7 @@ class TestPyMcdsOmeTiff(object):
 
     def test_mcds_make_ome_tiff_int(self, mcds=mcds):
         a_ometiff = mcds.make_ome_tiff(cell_attribute='cell_count_voxel', conc_cutoff={}, focus=None, file=False)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(a_ometiff)) == "<class 'numpy.ndarray'>") and \
               (a_ometiff.dtype == np.float32) and \
               (a_ometiff.shape == (4, 1, 200, 300)) and \
@@ -973,7 +973,7 @@ class TestPyMcdsOmeTiff(object):
 
     def test_mcds_make_ome_tiff_float(self, mcds=mcds):
         a_ometiff = mcds.make_ome_tiff(cell_attribute='pressure', conc_cutoff={}, focus=None, file=False)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(a_ometiff)) == "<class 'numpy.ndarray'>") and \
               (a_ometiff.dtype == np.float32) and \
               (a_ometiff.shape == (4, 1, 200, 300)) and\
@@ -986,7 +986,7 @@ class TestPyMcdsOmeTiff(object):
 
     def test_mcds_make_ome_tiff_conccutoff(self, mcds=mcds):
         a_ometiff = mcds.make_ome_tiff(cell_attribute='ID', conc_cutoff={'oxygen': -1}, focus=None, file=False)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(a_ometiff)) == "<class 'numpy.ndarray'>") and \
               (a_ometiff.dtype == np.float32) and \
               (a_ometiff.shape == (4, 1, 200, 300)) and \
@@ -999,7 +999,7 @@ class TestPyMcdsOmeTiff(object):
 
     def test_mcds_make_ome_tiff_focus(self, mcds=mcds):
         a_ometiff = mcds.make_ome_tiff(cell_attribute='ID', conc_cutoff={}, focus={'default'}, file=False)
-        assert(str(type(mcds)) == "<class 'pcdl.pyMCDS.pyMCDS'>") and \
+        assert(str(type(mcds)) == "<class 'pcdl.timestep.TimeStep'>") and \
               (str(type(a_ometiff)) == "<class 'numpy.ndarray'>") and \
               (a_ometiff.dtype == np.float32) and \
               (a_ometiff.shape == (1, 1, 200, 300)) and \
